Create new specimen controller in specimen module. Move some specimen actions and JSPs to study-api and specimen module.
