+++ api/src/org/labkey/api/study/SpecimenUrls.java
@@ -34,13 +34,14 @@ public interface SpecimenUrls extends UrlProvider
     ActionURL getCompleteSpecimenURL(Container c, String type);
     ActionURL getDeleteRequestURL(Container c, String id);
     ActionURL getInsertSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
+    ActionURL getManageActorsURL(Container c);
     ActionURL getManageRequestURL(Container c);
     ActionURL getManageRequestStatusURL(Container c, int requestId);
+    ActionURL getOverviewURL(Container c);
     ActionURL getRequestDetailsURL(Container c, String requestId);
     ActionURL getRequestDetailsURL(Container c, int requestId);
     ActionURL getShowCreateSpecimenRequestURL(Container c);
     ActionURL getShowGroupMembersURL(Container c, int rowId, @Nullable Integer locationId, ActionURL returnUrl);
-    ActionURL getShowSearchURL(Container c);
     ActionURL getSpecimenEventsURL(Container c, ActionURL returnURL);
     ActionURL getSpecimensURL(Container c);
     ActionURL getSpecimensURL(Container c, boolean showVials);
