+++ api/src/org/labkey/api/study/StudyUrls.java
@@ -19,7 +19,9 @@ import org.labkey.api.action.UrlProvider;
 import org.labkey.api.data.Container;
 import org.labkey.api.exp.api.ExpProtocol;
 import org.labkey.api.exp.api.ExpSampleType;
+import org.labkey.api.security.User;
 import org.labkey.api.view.ActionURL;
+import org.labkey.api.view.NavTree;
 import org.springframework.web.servlet.mvc.Controller;
 
 /**
@@ -44,4 +46,5 @@ public interface StudyUrls extends UrlProvider
     ActionURL getLinkToStudyURL(Container container, ExpProtocol protocol);
     ActionURL getLinkToStudyConfirmURL(Container container, ExpSampleType sampleType);
     ActionURL getLinkToStudyConfirmURL(Container container, ExpProtocol protocol);
+    void addManageStudyNavTrail(NavTree root, Container container, User user);
 }
