+++ study/api-src/org/labkey/api/specimen/actions/HiddenFormInputGenerator.java
@@ -1,0 +1,8 @@
+package org.labkey.api.specimen.actions;
+
+import org.labkey.api.view.ViewContext;
+
+public interface HiddenFormInputGenerator
+{
+    String getHiddenFormInputs(ViewContext ctx);
+}
