+++ study/src/org/labkey/study/controllers/StudyController.java
@@ -387,6 +387,11 @@ public class StudyController extends BaseStudyController
             return url;
         }
 
+        @Override
+        public void addManageStudyNavTrail(NavTree root, Container container, User user)
+        {
+            _addManageStudy(root, container, user);
+        }
     }
 
     public StudyController()
