+++ study/src/org/labkey/study/view/manageStudy.jsp
@@ -60,7 +60,7 @@
 <%@ page import="org.labkey.study.model.StudyImpl" %>
 <%@ page import="org.labkey.study.model.StudyManager" %>
 <%@ page import="org.labkey.study.model.StudySnapshot" %>
-<%@ page import="org.labkey.study.view.specimen.ManageSpecimenView" %>
+<%@ page import="org.labkey.api.specimen.view.ManageSpecimenView" %>
 <%@ page import="java.util.Collection" %>
 <%@ page import="java.util.LinkedList" %>
 <%@ page import="java.util.List" %>
